+++ schemas/lab_result.py
@@ -7,6 +7,6 @@ class Lab_Result(BaseModel):
     parameter_id : Optional[int] = None
     unit : Optional[str] = None
     result : Optional[str] = None
-    upper_limit : Optional[int] = None
-    lower_limit : Optional[int] = None
+    upper_limit : Optional[float] = None
+    lower_limit : Optional[float] = None
     remark : Optional[str] = None
